👌 IMPROVE: merged scripts to one file
