+++ src/djs/functions/create/createEmbed.js
@@ -28,7 +28,7 @@
  */
 
 const { EmbedBuilder } = require('discord.js');
-const { logError } = require('../scripts/logError.js');
+const scripts = require('../scripts/scripts.js');
 
 /* Example embed object that gets passed in below
 const embedObj = {
@@ -73,7 +73,7 @@ module.exports = function createEmbed(obj) { // DJS v14
   if (!obj.title && !obj.description && !obj.image && !obj.fields.length) {
     // If not, log an error
     try {
-      logError(new Error('Invalid properties were given to create the embed'), 'Invalid properties were given to create the embed');
+      scripts.logError(new Error('Invalid properties were given to create the embed'), 'Invalid properties were given to create the embed');
     } catch (error) {
       console.error(error);
     }
@@ -86,7 +86,7 @@ module.exports = function createEmbed(obj) { // DJS v14
     embed = new EmbedBuilder();
   } catch (error) {
     try {
-      logError(error, 'Error creating EmbedBuilder instance');
+      scripts.logError(error, 'Error creating EmbedBuilder instance');
     } catch (error) {
       console.error(error);
     }
@@ -104,7 +104,7 @@ module.exports = function createEmbed(obj) { // DJS v14
     if (obj.author) embed.setAuthor(obj.author.name, obj.author.iconURL, obj.author.url);
   } catch (error) {
     try {
-      logError(error, 'Error setting properties of the embed');
+      scripts.logError(error, 'Error setting properties of the embed');
     } catch (error) {
       console.error(error);
     }
@@ -120,7 +120,7 @@ module.exports = function createEmbed(obj) { // DJS v14
     }
   } catch (error) {
     try {
-      logError(error, 'Error adding fields to the embed');
+      scripts.logError(error, 'Error adding fields to the embed');
     } catch (error) {
       console.error(error);
     }
