--- src/djs/functions/scripts/verifyImport.js
@@ -1,8 +1,0 @@
-module.exports = function verifyImport(module) {
-    // check if the module is defined
-    if (module !== undefined) {
-      return true;
-    } else {
-      return false
-    }
-}
